Доработка тестов
