+++ src/test/java/ru/skypro/homework/controller/AdsControllerTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.security.Principal;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,8 +21,12 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.AdsDto;
 import ru.skypro.homework.dto.CreateAdsDto;
+import ru.skypro.homework.dto.TotalNumberAds;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.PictureService;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.validateMockitoUsage;
 
 @ContextConfiguration(classes = {AdsController.class})
 @ExtendWith(SpringExtension.class)
@@ -37,10 +42,10 @@ class AdsControllerTest {
 
     @Test
     void testAddAds() {
-        AdsController adsController = null;
-        CreateAdsDto properties = null;
-        MultipartFile image = null;
-        Principal principal = null;
+        AdsController adsController = mock(AdsController.class);
+        CreateAdsDto properties = mock(CreateAdsDto.class);
+        MultipartFile image = mock(MultipartFile.class);
+        Principal principal = mock(Principal.class);
 
         ResponseEntity<AdsDto> actualAddAdsResult = adsController.addAds(properties, image, principal);
     }
@@ -66,15 +71,7 @@ class AdsControllerTest {
 
         ResultActions actualPerformResult = buildResult.perform(requestBuilder);
     }
-    @Test
-    void testGetAdsMe() throws Exception {
-        Object[] uriVariables = new Object[]{};
-        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.get("/ads/me", uriVariables);
-        Object[] controllers = new Object[]{adsController};
-        MockMvc buildResult = MockMvcBuilders.standaloneSetup(controllers).build();
 
-        ResultActions actualPerformResult = buildResult.perform(requestBuilder);
-    }
     @Test
     void testGetAllAds() throws Exception {
         Object[] uriVariables = new Object[]{};
@@ -87,7 +84,7 @@ class AdsControllerTest {
 
     @Test
     void testGetImage() throws Exception {
-        Object[] uriVariables = new Object[]{"42"};
+        Object[] uriVariables = new Object[]{42};
         MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.get("/ads/image/{id}", uriVariables);
         Object[] controllers = new Object[]{adsController};
         MockMvc buildResult = MockMvcBuilders.standaloneSetup(controllers).build();
@@ -97,7 +94,7 @@ class AdsControllerTest {
     @Test
     void testRemoveAds() throws Exception {
         Object[] uriVariables = new Object[]{1};
-        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.delete("/ads/{id}", uriVariables);
+        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.delete("/ads/{id}", uriVariables, mock(Principal.class));
         Object[] controllers = new Object[]{adsController};
         MockMvc buildResult = MockMvcBuilders.standaloneSetup(controllers).build();
 
@@ -106,13 +103,13 @@ class AdsControllerTest {
     @Test
     void testUpdateAds() throws Exception {
         Object[] uriVariables = new Object[]{1};
-        MockHttpServletRequestBuilder contentTypeResult = MockMvcRequestBuilders.patch("/ads/{id}", uriVariables)
+        MockHttpServletRequestBuilder contentTypeResult = MockMvcRequestBuilders.patch("/ads/{id}", uriVariables, mock(Principal.class))
                 .contentType(MediaType.APPLICATION_JSON);
 
         CreateAdsDto createAdsDto = new CreateAdsDto();
         createAdsDto.setDescription("Характеристики");
-        createAdsDto.setPrice(1);
-        createAdsDto.setTitle("Dr");
+        createAdsDto.setPrice(1000);
+        createAdsDto.setTitle("Настольная лампа");
 
         ObjectMapper objectMapper = new ObjectMapper();
         MockHttpServletRequestBuilder requestBuilder = contentTypeResult
