--- src/test/java/ru/skypro/homework/controller/AuthControllerTest.java
@@ -1,4 +1,0 @@
-package ru.skypro.homework.controller;
-
-public class AuthControllerTest {
-}
