--- src/test/java/ru/skypro/homework/controller/CommentControllerTest.java
@@ -1,4 +1,0 @@
-package ru.skypro.homework.controller;
-
-public class CommentControllerTest {
-}
