--- src/test/java/ru/skypro/homework/service/CommentServiceImplTest.java
@@ -1,4 +1,0 @@
-package ru.skypro.homework.service;
-
-public class CommentServiceImplTest {
-}
