--- src/test/java/ru/skypro/homework/service/UserServiceImplTest.java
@@ -1,8 +1,0 @@
-package ru.skypro.homework.service;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public class UserServiceImplTest {
-}
